+++ printf/src/parse_unsigned_integers.c
@@ -6,7 +6,7 @@
 /*   By: skoskine <skoskine@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/20 09:06:28 by skoskine          #+#    #+#             */
-/*   Updated: 2021/04/09 09:37:18 by skoskine         ###   ########.fr       */
+/*   Updated: 2021/04/09 10:20:54 by skoskine         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -48,7 +48,7 @@ static uintmax_t	get_unsigned_arg(t_data *specs, va_list *ap)
 
 static int	update_uint_specs(t_data *specs, uintmax_t value, char *value_str)
 {
-	int	len;
+	size_t	len;
 
 	if (value == 0)
 		specs->is_zero = 1;
